\name{CTT-package}
\alias{CTT-package}
\alias{CTT}
\docType{package}
\title{
Classical Test Theory Functions
}
\description{
This package can be used to perform a variety of tasks and analyses associated with classical test theory (CTT): 
score multiple-choice responses, perform reliability analyses, conduct item analyses, and transform scores into 
different scales}
\details{
\tabular{ll}{
Package: \tab CTT\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2008-06-08\cr
License: \tab GPL version 2 or newer\cr
}
Functions the "CTT" package has:
reliability; 
score;
distractor.analysis;
score.transform;
spearman.brown;
subscales;
}
\author{
John Willse, Zhan Shu
}
\Maintainer: John Willse<willse@uncg.edu>
}
\references{
     Crocker, L. & Algina, J. (1986). Introduction to Classical & Modern Test Theory, New York: Harcourt Brace Jovanovich College Publishers.
     Cronbach, L. J. (1951). Coefficient alpha and the internal structure of tests. Psychometika, 16, 297-334.
     Gulliksen, H. (1950). Theory of Mental Tests. New York: John Wiley & Sons, Inc.

}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }

